+++ src/texture_sample.cpp
@@ -8,8 +8,8 @@ ScopedRenderTexture TextureSample::down_sample(const Texture& source, const floa
     const RenderTexture render_texture = LoadRenderTexture(source.width / factor, source.height / factor);
 
     BeginTextureMode(render_texture);
-        const Rectangle source_rect = Rectangle { 0, 0, source.width, source.height };
-        const Rectangle destination = Rectangle { 0, 0, render_texture.texture.width, render_texture.texture.height };
+        const Rectangle source_rect = Rectangle{ 0, 0, (float)source.width, (float)source.height };
+        const Rectangle destination = Rectangle{ 0, 0, (float)render_texture.texture.width, (float)render_texture.texture.height };
 
         SetTextureFilter(source, TEXTURE_FILTER_BILINEAR);
         DrawTexturePro(source, source_rect, destination, {}, 0, WHITE);
